+++ subprojects/messaging/src/main/java/org/gradle/internal/serialize/kryo/StringDeduplicatingKryoBackedEncoder.java
@@ -99,7 +99,7 @@ public class StringDeduplicatingKryoBackedEncoder extends AbstractEncoder implem
     /**
      * Returns the total number of bytes written by this encoder, some of which may still be buffered.
      */
-    public int getWritePosition() {
+    public long getWritePosition() {
         return output.total();
     }
